paths: Add unit tests for existing path system
The existing path system hasn't ever had any unit tests, and has been
tested only very lightly in the Halyard Test project.  This patch adds
some unit tests so we can verify that the old path system works, and
make sure that when we implement the new path system, we only change
it in ways that we intend.