Test fix: Restructure CommonTest to always have a TSchemeInterpreter
Commit 42bcc7dffe9bcbadc1ce482807cec035ade54404 broke the command-line test
suites.  (I only tested it with the GUI-based test suites, unfortunately.)
As it turns out, the test cases in TSchemeConv.cpp have always assumed that
they could access the sScriptEnv of a destroyed TInterpreter object.  When
we made sScriptEnv into a member variable, we forced everybody to access it
through TSchemeInterpreter::GetSchemeInterpreter(), which raises an
assertion if the interpreter has been destroyed.

Of course, this odd behavior was only possible because of the odd relation
between CommonTest and TSchemeInterpreterTests.cpp.  Specifically,
TSchemeInterpreterTests.cpp created and destoyed a TSchemeInterpreter
object, and CommonTest relied on other unit tests being able to poke
through the remnants of that TSchemeInterpreter's static variables.  Yuck.

The fix: Have CommonTest create a single, top-level TSchemeInterpreter and
keep it around for any test suites that want to use it.  This requires us
to special-case the handling of TSchemeInterpreterTests.cpp, which actually
runs most of its unit tests before calling a new RunAllCppTests primitive.
The RunAllCppTests primitive then runs the rest of our old-style and
new-style test cases.

I've done quite a bit of hand-testing for this commit:

 * I manually caused test failures and threw exceptions from Scheme
   interpreter tests, old-style tests, and new-style tests.  In each case,
   the errors were detected, though they sometimes resulted in a Win32
   dialog that blocked further execution of CommonTest.
 * I ran the GUI-based C++ test harness.
 * I ran all the cards in "tests/".
 * I opened the GUI editor and poked at it briefly to make sure that it
   didn't have any immediately obvious problems with the recent work in
   Common/lang/scheme.

[Merge warning: I had to rename (test ...) to (check ...) in
TestScripts/start.ss to avoid a conflict with mizzen-unit's (test ...)
macro.]