wxMac: Compile DrawingArea
This was simply a matter of conditionalizing all the Quake-specific code
using CONFIG_HAVE_QUAKE2.  Unfortunately, this makes the code rather ugly.
We need to factor out these #if / #endif blocks into individual functions,
as per our standard coding style.  But to do that, I'll need to be running
on Windows machine.