wxMac: Compile TWxPrimitives
This involved several major changes:

 1) Conditionalizing ActiveX and AudioStream support.
 2) Adding ToWxString to CommonWxConv.
 3) Refactoring most of the FIND_ELEMENT macro into a find_element<T>
    function.

Note that, if we want, we can go back through the source code and grep for
wxConvLocal, and use the new ToWxString function instead.  We might also
want to add a ToStdString function.  These changes would probably make the
code cleaner.

We've temporarily skipped over one issue with this file: It still refers to
BrowserElementIE, which has a portable header but a non-portable *.cpp
file.  We'll get to this shortly.