wxMac: Conditionalize FancyCrashReporter and BrowserElementIE
FancyCrashReporter currently relies on a bunch of wxWidgets 2.6 APIs.
We'll look at it later.

BrowserElementIE has been conditionalized in a fashion similar to
MovieWindowQT.  I'm not sure this is a good way to conditionally build
different versions of a class, but at least it's consistent with what we're
already doing.