wxMac: It compiles! (But it doesn't run.)
The EventDispatcher timestamp support (which detects stale events) has been
disabled, because none of wxMac backends appears to set up timestamps.
We'll want to investigate this further at some point.

When run, Halyard pauses for several seconds and then crashes with a bus
error.  Since we've removed it's event loop, this is to be expected.