wxMac: Preliminary sketch of event loop
Halyard requires a custom event loop.  This event loop is run under control
of the Idle primivitive, which is called by the interpreter when it wants
to yield time to the operating system.

This is a preliminary implementation of Halyard's event loop for wxWidgets
2.8.  It assumes that we're running on one of the platforms using
wxEventLoopManual, which has a Run method roughly similar to ours.