wxMac: Don't set Stage background before Create()
We used to set up the background-drawing parameters for Stage, *then* call
Create(), to avoid early repainting of this window with the wrong options.
But that doesn't work on wxMac, and the application hangs at startup.  So
we've re-ordered the SetBackground* operations after Create(), and we must
now rely on StageFrame to not Show() us until we're ready.

Current status: The StageFrame and StartupDlg get shown, and then Halyard
hangs, with no icon in the task bar, and no event processing other than the
close button in the StartupDlg title bar.