wxMac: Fixes to compile against Leopard-bundled wxWidgets
I am building Halyard against the wxWidgets provided with Leopard, rather
than the one from MacPorts, and there are a few incompatibilities.

The most common fix is that there are several strings that need to be
wrapped in a wxT(...).  There are also two cases of mb_str() used in
places it doesn't work; in one case, it was used somewhere expecting a
std::string, and in the other, it was used in a place expecting a char
*.  It seems that in some configurations, mb_str() returns a char *,
which can be implicitly cast to a std::string, so everything just
works, but in other configurations, it returns a wxCharBuffer, which
needs to be explicitly converted.

[It looks like MacPorts wxWidgets 2.8.7 includes a few extra Unicode
conversion methods for things like wxLogTrace.  I expect Brian will have a
few more of these shortly. -Eric]