paths: Fix .resolve-path errors to actually mention the path name
We had been using a default value for the keyword argument
:if-not-found to .resolve-path, that gave an error message referencing
SELF.  This didn't give useful error messages, since in the context of
the parameter list, the SELF in scope is the class, not the instance.
This patch fixes it to have the default error message code inside the
body of .resolve-path, so it actually picks up the correct name, as
well as having a useful distinction in the error message between
absolute and relative paths.