wxMac: Add Xcode project for debugging Halyard
This adds a bare-bones Xcode project that just runs 'make' to build
Halyard, so we can more easily rebuild Halyard and debug it while
continuing to use our current build system.

It appears that you may have to add Halyard.app as a custom executable
your self, since that seems to go in the per-user project files as
opposed to the generic project file.