Update Makefiles to support building in separate directories
This patch allows us to run builds from within different directories,
to support separate debug and release builds or other strange build
settings that we don't want to interfere with our normal build.

To do a build from a separate directory:
  $ mkdir debug
  $ cd debug
  $ ../configure
  $ make

Note: This seems to break pre-compiled headers.  We'll want to try
and fix that later if possible.

Paired-with: Eric Kidd <eric.m.kidd@dartmouth.edu>