wxMac: Fix assertion failure when creating 0-area DrawingArea
If you run wxWidgets with assertions turned on in the Mac port, you
will get an assertion failure upon jumping to the /index card.  This
is because the /index card has a 0-width overlay, which triggers a bug
in wxWidgets.  Rather than waiting for the bug to be fixed upstream,
this patch simply doesn't create any bitmaps with area 0, and ensures
that nothing tries to access the bitmap when the DrawingArea has area
0 by calling HasAreaOfZero().

v2: Remove spurious difference in inHasAlpha conditional