wxMac: Fix assertion failure when drawing elements partly off-screen
In the Mac build of the engine, with wxWidgets in debug mode, we would
get an assertion failure on the /features/mask/light card if we moved
the cursor to the left or top edge of the screen.  This is because a
rectangle that went off the edge of the screen would be marked dirty,
and later when we tried to blit that, we'd get an assertion failure
because we were trying to blit from rects that were not contained in
the compositing pixmap.

This patch fixes that by intersecting the dirty region with the screen
rect before trying to blit.