Allow loader.ss to use (require ...) paths
This requires us to set up our collection paths before we actually define a
module in loader.ss.  But because top-level code in loader.ss will actually
be run _before_ we have a TSchemeInterpreter, change how we handle
SetCollectionDir.

I'm not quite sure why this worked before, but if I had to guess, I'd say
the PLT actually delays module initialization until you try to use a
module.  And in our case, we didn't try to use anything in loader.ss until
after TSchemeInterpreter was initialized.