Modify loader to use sandbox.ss instead of custom code
The goal: We want overhaul our compilation code to work in a much more
sensible fashion.  And we also want to enable errortrace.ss, which will
give us stack traces.  But our custom loader is a bit hairy, and it has
been evolving over many years.

Eli Barzilay, one of the PLT Scheme maintainers, has suggested that we
convert from our custom loader code to something based on sandbox.ss, a
module in PLT's mzlib.  This module is intended for running student
programs and other untrusted source code in mzscheme applications.

Unfortunately, sandbox.ss uses mzscheme threads.  And mzscheme threads have
a number of important limitations:

http://download.plt-scheme.org/doc/372/html/insidemz/insidemz-Z-H-8.html#node_sec_8.1

This is a huge patch.  Here's an overview of what's going on:

 1) halyard/loader.ss has been split into two files:
    halyard/loader/stage1.ss and halyard/loader/stage2.ss.  The stage1.ss
    file runs inside the global environment, and it is responsible for
    creating the sandboxed evaluator thread.  The stage2.ss file now runs
    entirely within the sandbox, and it is in charge of setting up the
    compiler system and actually loading the script.

 2) HALYARD_BEGIN_STACK_BASE and HALYARD_END_STACK_BASE have been moved
    down into TInterpreterManager::Run.  This minimizes the number of stack
    frames affected by stack-relocating context switches.  (Actually, I'm
    not sure if mzscheme actually relocates stack frames for threads.  But
    it does it for a variety of other reasons, including continuations, so
    better safe than sorry.)  This has necessitated a number of
    restructurings elsewhere in the code, including delayed initialization
    of TSchemeInterpreterManager state.  See MZSCHEME-THREADS.txt for
    details of what's going on here.

 3) Lots of state-tracking variables and ASSERT calls have been added, to
    make sure that we're obeying all the new rules.

v2:
  Cherry-picked from old sandbox branch
  Merged 22d4b19dd2c48961c9aa9b7b8e53d4e40a4568b4 manually