Properly convert from wxChar * to char * in AppLog
As part of the Mac port, we've been using the Unicode build of
wxWidgets, and adding all of the necessary conversions to strings to
support that.  Most of these have been found by the typechecker, but
in the case of printf type functions that take varargs, passing in the
wrong type of string is not caught. In AppLog, we had been passed a
wxChar *, and were passing it along to our logging commands.  Instead,
we now convert it (implicitly) to a wxString, and convert that to a
std::string, and then use .c_str() to extract a char * that can be
printed using printf-type functions.