wxMac: Update the script editor to work with vanilla wxWidgets 2.8
In order to complete the upgrade from wxWidgets 2.6 to the latest
release of wxWidgets 2.8, it is necessary to remove its dependencies
on our patched version of wxStyledTextControl and update it to build
against Unicode builds of wxWidgets.

This patch achieves that goal by building ScriptEditor on the Mac
against vanilla MacPorts wxWidgets 2.8.7.  It updates all of the string
handling to use conversion functions that will allow it to build on the
Mac; code review would be helpful on the ones that use wxString::Printf,
since it was unclear how to properly pass strings into the varargs
arguments.

This patch also introduces the CONFIG_HAVE_CUSTOM_STC macro which is
used to enable or disable features depending on whether we're building
against our custom wxStyledTextControl (with our patched LexLisp and
find functionality) or the default one.  We are hoping to get the find
functionality pushed upstream, and to fork LexLisp into our own LexHalyard
or LexMizzen that is built separately from stc, but for now, we'd like
to run even with the vanilla stc.  We do this by disabling the find
functionality in the vanilla version, and merging all of the different
keyword classes passed into the lexer into the two provided by vanilla
LexLisp, using one for macros and one for everything else.

Known issues: For some reason, certain menu items (About, Show
Listener) aren't switched when you switch between the windows, and so
they act like no-ops while passing the unmodified key into Scintilla.
Also, there appear to be lots of things broken with using XPMs for
wxBitmaps; if you build against the debug version of wxWidgets, you
get all kinds of assertion failures, many of the icons display
incorrectly, and some don't display at all.