Fix browser test cases that fail on wxMac
We use wxWidgets fallback browser on the Mac, which causes some of our
browser test cases to fail.  In particular, 'about:blank' is not a
valid URL in the fallback browser, and pages that return 404 errors
cause WX errors.  This patch changes %browser% to not have a default
URL value (so it must always be specified when constructing a
browser), and changes the tests to all be valid given that change.