Export %pref and set-%pref! from data-file.ss
There is apparently a bug in the module system that prevents
define-pref from expanding properly unless the %pref and set-%pref!
functions that it expands to are also exported.  We generally
deprecate define-pref, but VTRA uses it, so we're exporting them
to make porting VTRA to the new engines easier.

Bug-ID: 5137