wxMac: Fix crash in mTree->SetItemBold
As noted in 22d4b19dd2c48961c9aa9b7b8e53d4e40a4568b4, we've been seeing an
intermittent crash in ProgramTree::NotifyEnterCard on the Mac.  This
occurred here:

    // Move the highlighting to the appropriate card.
    if (mHaveLastHighlightedItem)
        mTree->SetItemBold(mLastHighlightedItem, FALSE);

We were forgetting to reset mHaveLastHighlightedItem when the script was
reloaded, causing us to refer to a deallocated mLastHighlightedItem.  The
first time we did this (when we entered /start), we actually got away with
it.  But when we entered the card /index immediately afterward, we would
occasionally run into corrupted state and crash.

I tracked this down because I feared it was a GC bug (and the last commit
made it trigger much more often), but it turned out to be a (presumably
cross-plaform) memory corruption issue instead.