Replace wrapped load/use-compiled-handler with callback
In the past, we've relied on wrapping the return value of
(make-compilation-manager-load/use-compiled-handler) with a function that
called (heartbeat) and (update-splash-screen!).  But this required us to
use a patched compilation manager, because the standard
(make-compilation-manager-load/use-compiled-handler) refuses to be
overridden:

  5d9327d487a08c909d271513fa314e7f82eb3aba

Here, we simply hook (manager-compile-notify-handler) with our heartbeat
function.  As a side effect, this enables script compilation on the
Macintosh.

We also add some more tracing code to help debug compilation manager
problems in the future.