Show element names in status bar when borders are displayed
Turn on "Display Element Borders", and this patch will show the name of the
current element in the status bar.  This is a little trickier than I had
anticipated, because we also want to show the names of elements that would
otherwise be invisible to the mouse.

Note that we're a bit sloppy about cleaning up our internal state when this
option is turned off.  This doesn't present any correctness issues for
Halyard itself, but it may occasionally lead to stale displays until the
next time the status bar gets updated.  This slight sloppiness makes the
code much simpler.

Note that we also modify %text-box%'s :clickable-where-transparent?
property to default to #t, because nobody wants to try to hover the mouse
over individual pixels in a letter.