Editor: Save open tabs and currently active tab
This patch adds tab saving to the editor.  We save the currently
open tags and active tab every time a tab is opened, closed, or
highlighted, so if the engine crashes or needs to be quit, the
scriptor can get most state back when re-launching.

In order to do this, we introduce the notion of a
project-specific config file that is local to a given developer.
This is a file named project.config in the Halyard program
directory, alongside the application.halyard file.  It is in INI
file format, as provided by a wxFileConfig object owned by
HalyardApp.  We read the config in as soon as we have setup our
script direcotry, and thereafter we can call
wxGetApp().GetProjectConfig() to fetch the config object, and
wxGetApp().SaveProjectConfig() to write it back out.