Compile all *.ss files to *.zo in a single pass
The problem: The first compile of halyard/test took ~26 seconds, the second
took roughly ~8 seconds, and all following compiles took roughly ~2
seconds.

What's going on: The file (lib "cm-ctime.ss" "mzlib" "private") is only
loaded by the compilation manager when it actually compiles something.
Unfortunately, this messes up the progress bar calculations, because we
never know how many files we need to load.  In a misguided attempt to work
around this problem, we were loading this file from kernel.ss, to make sure
it always got loaded.

Unfortunately, this confused the compilation manager, and caused it to
recompile all files starting with kernel.ss on the _second_ reload
(starting from a clean system).

So now we're trying another strategy for including this file.