Fix out-of-memory crash during script reload
Previously, Halyard would run out of memory once the Scheme heap exceeded
~400MB.  This tended to make large scripts crash during the compilation
process, especially when errortrace was turned on.

At Matthew Flatt's suggestion, we bumped LOG_APAGE_SIZE from 14 to 16.
This should theoretically quadruple the amount of RAM we can allocate
before the engine crashes with a NULL pointer write in GC_mark (which is
caused by a VirtualAlloc failure).