Store identifier and indentation information in ScriptEditorDB
Up until now, the ScriptEditor has relied on two sources of information
about the program:

  1) TInterpreter::GetAllIdentifiers, which fetched a list of identifiers
     from the running program, and used it for syntax highlighting,
     identifier completion, and indentation.
  2) ScriptEditorDB, which extracted information directly from the source
     code, and used it for meta-dot support and tooltips.

In the original editor, we only used (1).  We added (2) at a later date.
But (1) has always had lots of problems--it didn't work until a script was
fully loaded, it couldn't complete identifiers that weren't exported to the
top level, and it wasn't going to be compatible with our designs for
demand-loading of source code.

This patch keeps the original source-scanning machinery from (1), but we no
longer run it on the top-level namespace.  Instead, we run it only on the
mzscheme module, which has no on-disk source files.  And instead of passing
this information to the ScriptEditor, we instead store it in the
ScriptEditorDB, from which it can be fetched by the ScriptEditor as needed.

The required adding a new database table for indentation information, and
replacing the old indentation machinery with something based on
define-syntax-tagger*.  We also moved indent.ss from mizzen to halyard,
because it's actually Halyard-specific functionality.

There's a few snippets of code here that may not be refactored in a
completely ideal fashion.  In particular, we don't get the
language-dependent vs. language independent boundry 100% correct.  These
places are commented in the source.

Also worth noting: Now that we have one "fake" file in the ScriptEditorDB,
the ScriptTree class needs to skip that file when building the tree.