Added begin/var support to λ special form
Just in case anybody wants to use the new λ special form with Halyard
(which probably isn't a good idea for now, although we want full Unicode
support later on), we replace the mzscheme version with a version aware of
begin/var.

The only real justitication for doing this is to start flushing out Unicode
bugs (and to amuse me, but that's not a particularly good justification by
itself).