reorg: Move Runtime/, Fonts/ to new runtime/ directory
Our goal: Make the engine look for the Runtime/ and Fonts/ directories
relative to the engine, not relative to the current script.

The basic strategy: We move Runtime and Fonts to a new top-level runtime/
directory, which (on Windows) we also use as the output directory for
building our *.exe and *.dll files.

  test/Runtime      -> runtime/collects
  test/Fonts        -> runtime/fonts
  Win32/Bin/*       -> runtime/
  test/test-scripts -> Common/test/scripts
  test/fixtures     -> Comment/test/fixtures

Newly-introduced problems:
  1) The SetScriptsDirectoryName API is now just silly and redundant.
  2) The ScriptEditorDB doesn't know how to scan the new runtime/ directory.
  3) The Mac build is broken in a variety of exciting ways.
  4) We haven't updated the scripts which release engines.

Older problems which affect us:
  5) Halyard can't operate in Runtime mode because the user.conf code tries
     to create the temporary directory even when the editor isn't being run.
     This is a pre-existing bug that we need to fix fairly urgently.