reorg: Fix ScriptEditorDB to scan runtime/ directory
Now that some of our *.ss files actually live in the engine's
runtime directory, and not in the script's root directory, we need to
make the ScriptEditorDB a bit smarter.

Unfortunately, the ScriptEditorDB relies heavily on "relpaths",
which are paths relative to the script's root directory.  To
support the new layout, we define a new kind of semi-magic
relpath.  Specifically, relpaths of the form "runtime/foo" are
treated as abbreviations for $RUNTIME_DIR/foo instead of
$SCRIPT_ROOT/runtime/foo.

Some other changes worth mentioning:
  1) We now allow directories such as $SCRIPT_ROOT/collects to be missing
     without raising any errors.  This makes things easier for git users,
     who can't check empty directories into version control.
  2) The file TestAll.cpp no longer specifies a runtime directory
     of "../../runtime", because that made it hard to compare
     paths in test suites.  Instead, we use a new
     Path::ParentDirectory method to strip two directories off
     the current path manually.
  3) Conversion between boost::filesystem::path and FileSystem::Path is
     now much more systematic.
  4) NormalizePath has been renamed to CompletePath, which is a
     more accurate description.
  5) The ScriptEditorDB is now much more reliable about calling
     RelPathToNative when necessary.
  6) ScriptEditorDB::ProcessFileInternal now takes a native path, not a
     relpath.

Reviewed-by: Brian Campbell <brian.p.campbell@dartmouth.edu>
Fix-for: ce4f4aca4860360f7de2d5eb1519010f1d8a58b3