reorg: Mac: Add support for new runtime/ directory location
The Mac build of the reorg branch has bit-rotted a bit since yesterday.
Here, we make a number of simple build system fixes, and also address the
big problem: Finding the runtime directory.

We symlink runtime/ to wx/src/Halyard.app/Contents/SharedSupport, which is
the MacOS equivalent of /usr/share/halyard--the spot where unchanging
support files are expected to be found.

This requires a small tweak to the ScriptEditorDB.  We teach it to follow
top-level symlinks in ScanTree.

Fix-for: ce4f4aca4860360f7de2d5eb1519010f1d8a58b3