Add syntax tagger for DEFPRIMCLASS, and fix tagger for DEFSUBST
In order to get proper auto-completion and highlighting for Swindle
classes representing primitive types, we need to add a syntax tagger
for DEFPRIMCLASS.  Doing this alone is insufficient, however, because
DEFSUBST was broken in certain usages (when defining multiple
substitutions at once), and when a syntax tagger can't match a form,
it errors out and stops processing that file, which was causing it to
miss all of the DEFPRIMCLASS forms.

This patch does not actually fix DEFSUBST to work completely correctly,
which would involve adding help strings for every possible pattern, but
DEFSUBST is only used like this in a few Swindle-internal places, so we
don't really care.