editor: Fix crash when file in saved active tab no longer exists
The new tab-saving logic in the script editor had a bug where if the
filename saved as the currently active tab no longer corresponded to
an existingg file, the editor would throw errors when you opened it to
restore tab state.

This happened because while we checked if files existed before opening
each tab, we didn't check if a file existed when restoring the active
tab.  The fix pushes the file existence check down into the
OpenDocumentInternal, so that anyone who tries to open a non-existant
file will instead just do nothing (which also occurred when you deleted
a file and then tried to get to it via meta-. or the file browser).