Pass correct parameters in RECTANGLE-OUTLINE node helper
When we implented the various node helpers, we neglected to test the
RECTANGLE-OUTLINE helper, and got the arguments to it wrong.  This
patch adds test cases for all of the node-helpers defined in
elements.ss, as well as all of the new-element forms other than the
media, for which the official tests are simply the cards demonstrating
them.