reorg: Allow content to be stored in collects/*/_halyard directories
We overhaul the content-finding system, and replace it a new
function: resolve-content-path.  This new function can search for
content in a variety of places, including:

  foo.ogg           streaming/media/foo.ogg (if present)
  foo.ogg           local/media/foo.ogg
  $COLLECT/foo.ogg  collects/$COLLECT/_halyard/local/media/foo.ogg

There's still a few fallback special cases for QuickTime media files, but
most of the rest of the content path-resolution code has been centralized.

We'll need to test the AudioStream and QuickTime network support
carefully.  It's possible that we'll find a regression or two.

Other changes:
  - ABSTRACT-PATH->NATIVE-PATH has gone away.  Use BUILD-PATH instead.
  - The new RESOLVE-CONTENT-PATH function is now an official public
    API, as are a variety of other useful things in
    runtime/collects/halyard/content-paths.ss.