Demand loading: Register groups with engine, not just cards
To support demand loading, we need to be able see groups that we haven't
actually loaded yet.  But this requires us to actually register groups with
the GUI, and not just infer their presence from the cards they contain.  We
do this by register all instances of %group-member%, the superclass of
%group% and %card%.

This also allows us to see empty groups in the GUI.

This is a pretty straight-forward change except for ProgramTree.cpp, which
was ancient and required a major overhaul:

 - Splitting of node names is no longer woven into the node-registration
   code.  It is now performed by AnalyzeNodeName.
 - The code for finding and creating both cards and groups has been merged
   into a single function.