Simplify Widget::Show error-handling further
Key insight: If HasVisibleRepresentation() is false, then IsShown() should
also be false.  So we can just use the inherited IsShown() method when
HasVisibleRepresentation() is true, and we shouldn't change the observable
behavior of Widget::Show.

This means that we don't have to defer further cleanup work until our next
development branch.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?8193