Demand loading: Add *enable-demand-loading?* variable
When this variable is set to #t, the engine behaves as always.  When this
variable is set to #f, then we refrain from loading code referred to by
EXTERNAL-GROUP and EXTERNAL-CARD.  Unfortunately, we have no ability to
install demand-loading proxies at the moment, so there's not yet any way to
load this code later.

To make this change, we needed to move DEFINE/P and some support code from
api.ss to kernel.ss.  With a few more small changes (described in
kernel.ss), we could actually move all of this into a separate file.