Demand loading: Change register-group-member parameters for clarity
Up until now, we sometimes referred to unloaded nodes as "placeholders" (in
the node registration machinery and in C++), and sometimes referred to them
as "trampolines" (in the higher-level code of nodes.ss and
external-nodes.ss).

To fix this inconsistent terminology, the "placeholder?" flag on
register-group-member (and its implementation) has been inverted to become
a "loaded?" flag.  This better represents the issues that C++ cares about,
and it dodges the terminology issue entirely.