Demand loading: Make find-static-node work on external-group children
The following code should now work:

  (@/features/external-card .jump)

Here, both /features and /features/external-card need to be loaded from
disk, but the static-node-table only contains a trampoline for @/features.
This requires some slightly tricky code that actually recurses through
analyze-node-name, which (in turn) now has a new parameter to specify what
should happen when a parent node can't be found.

Most of the trickiness is actually required to handle nodes which _don't_
exist.  For example:

  ;; On card /index.
  (@media/qt/movies .jump)

This will first look for /index/media/qt/movies, and find-static-node must
fail by returning #f.  Once that occurs, the path code the tries to resolve
to /media/qt/movies, which succeeds.