Demand loading: Show "open folder" controls for unloaded groups
We now show an "open folder" control (a triangle on the Macintosh, or a
"[+]" on Windows) on groups which haven't been loaded yet.  Once the group
has been loaded, we check again whether the group has any children, and if
it doesn't, we hide the "open folder" control.

Next: Using the open folder control to trigger the loading of a group.