Walk up directory hierarchy to guess at appropriate current directory
When we moved our engine into engine/win32, we lost the ability to
pick up our developer.prefs file in relation to the directory the
engine starts in, as well as our ability to guess at the directory we
should default the Open dialog to.  This is because the current
directory doesn't actually get initialized until we choose an
application.halyard to open, by which time we've already failed to
find our developer.prefs and thus initialized the preferences
incorrectly.

In order to fix this, we try walking up the directory hierarchy,
attempting to find an application.halyard.  If we do, we set that as a
temporary working directory, so we can pick up config/developer.prefs
from that directory, and we can use that as our default when opening
the program.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?10267