Demand loading: Fix crash when %kernel-valid-card? triggers load
Here's another kernel entry point which can trigger demand-loading, and
which needs to be run from inside a callback context.

Most of this patch is actually a framework for automatically detecting bugs
of this sort, and for display detailed information on how to fix those
bugs.  Unfortunately, this means that too much code is starting to
accumulate in kernel.ss again.

Fix-for: db587ed90c28cb228d9c0bf4544630f4628aec53