Unit test framework: Display errors in order they occur
We've had some problems where because of a failure during setup-test that
caused future setup-test invocations to fail, and because errors were
displayed in the reverse order that they occurred, we could never see the
error that actually caused the tests to fail; it was displayed off the
bottom of the screen.

This patch collects the test failures in order, so when we display them,
if there are any problems that propagate from one test to another, we will
see the original error first.