Increase LOG_APAGE_SIZE to 17 for Windows XP
Apparently, a LOG_APAGE_SIZE of 16 gives us a maximum heap of around 1 GB
on Vista, but only ~450-550 MB on some XP systems.  This is not surprising,
because the ultimate limit on heap size is LOG_APAGE_SIZE times the number
of pages we can allocate, and the latter may vary between operating
systems.

Increasing the LOG_APAGE_SIZE to 18 causes Halyard to crash as soon as the
Scheme interpreter is initialized on my XP system.  So I'm going with 17 on
an experimental basis.  We'll obviously want to test this on Vista.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?5896