Implement simpe static node path syntax
Our path system does not deal very well with the distinction between
static nodes and running nodes; a path object usually acts as a proxy
for the running node, but occasionaly as a proxy for the static node.
This is confusing, and means that we need to do things like resolving
paths before querying them about various types of information.

Without breaking backwards compatibility, this patch introduces a new
syntax for explicitly referring to a static node.  This expands a path
like /foo/bar into a call to (find-static-node-or-error '/foo/bar).
This means that the path will be resolved immediately as soon as the
code is executed, and will return the static node itself, not a path
pointing to it.  This may cause portions of the chain leading to that
path to be lazy loaded, but the node itself may still be a trampoline
object.

This patch also introduces the predicate (static-node? obj) to determine
if an object is a static node, and rewrites the existing path expander to
use regular expressions instead of manual parsing.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?10472