Buildbot support: Add SUCCESS? parameter to exit-script
We want to be able to run Halyard under the control of Buildbot (and other
command-line tools).  To do this, we must improve Halyard's command-line
interface.

The first step is modifying (exit-script) to return an error code to the
shell.  We add a new, optional SUCCESS? parameter.  If passed, Halyard will
exit with a non-zero status, which traditionally indicates an error.

  $ ../runtime/Halyard_d.exe -e "(exit-script)" .; echo $?
  0
  $ ../runtime/Halyard_d.exe -e "(exit-script #f)" .; echo $?
  1

This code piggy-backs on some pre-existing machinery in TInterpreterManager
for returning a non-zero exit status.