Buildbot support: Make errors accessible in COMMAND_LINE mode
When Halyard is run in COMMAND_LINE mode, we need some way to get at the
error-messages from scripts.  Normally, we could just modify TLogger to
print the errors to std::cerr, but this won't work on Windows.  So on
platforms without a usable std::cerr, we create "temp/output.txt" and write
our errors there.

This patch has a number of pieces:

  1) A TLogger::SetIsStandardErrorAvailable function, which can be used by
     the GUI to let TLogger know about the current platform's capabilities.
  2) A TLogger::GetErrorOutput function, which returns a std::ostream
     pointing to either the console or to "temp/output.txt".
  3) Support for redirecting error dialogs to the console in COMMAND_LINE
     mode.
  4) Ruby and rake support for running Halyard and displaying
     "temp/output.txt".

We also set TestAll.cpp to COMMAND_LINE mode, which will cause it to exit
immediately when an error occurs, and not just hang around doing nothing.